net/mlx5e: Receive flow steering framework for accelerated Ethernet flows
Add the steering structure in between the outer RSS and the outer RSS's TIRs
to do additional offload operation such as IPsec crypto offload. Since
IPsec is the first application that uses this framework, the following
explanation are based on IPsec example.

By default, if the adapters support IPsec crypto offload, empty flow
table is created (0x40005) with one empty Flow group (0xe) and one
fixed Flow group (0xf) with a fixed FTE that forwards all ESP traffic
to the ipv4 ESP's RSS TIR (0x06). This steering framework ensures that other types of
traffics (such as TCP, GRE, UDP, ...) do not see additional hops.

|-- FTE: 0x6 (FWD) to (FLOW_TABLE:0x40005) ip_ver:0x4 ip_prot:ESP
|   +-- FT: 0x40005 (level: 0x34, type: NIC_RX)
|       |-- FG: 0xe (OUT_HDR MISC: |ethtype|frag|ip_prot|src_ip_31_0|dst_ip_31_0|)
|       +-- FG: 0xf (NO_MATCH)
|           +-- FTE: 0xffff (FWD) to (TIR:0x6)

When IPsec security association is added, a new IPsec rule will be added
in to the Flow group 0xe. The new IPsec steering rule are forwarded to
the IPsec error checking Flow table (0x03) before forwarding to the same
ipv4 ESP's RSS TIR (0x06). Note that the FLow table 0x03 implementation
is specific for IPsec but the general idea of this framework is that the
final destination will be the application's RSS TIR.

|-- FTE: 0x6 (FWD) to (FLOW_TABLE:0x40005) ip_ver:0x4 ip_prot:ESP
|   +-- FT: 0x40005 (level: 0x34, type: NIC_RX)
|       |-- FG: 0xe (OUT_HDR MISC: |ethtype|frag|ip_prot|src_ip_31_0|dst_ip_31_0|)
|       |   +-- FTE: 0x0 (FWD MODIFY_HDR) to (FLOW_TABLE:0x3) modify_header_id:0x246 ethtype:IPv4 ip_prot:ESP src_ip_31_0:192.168.1.65 dst_ip_31_0:192.168.1.64
|       |       +-- FT: 0x3 (level: 0x35, type: NIC_RX)
|       |           +-- FG: 0x12 (NO_MATCH)
|       |               +-- FTE: 0x0 (FWD MODIFY_HDR) to (TIR:0x6) modify_header_id:0x244
|       +-- FG: 0xf (NO_MATCH)
|           +-- FTE: 0xffff (FWD) to (TIR:0x6)

Issue: 1821406
Change-Id: I07bd7824cd925c5fbe36085851f9c483013dead0
Signed-off-by: Boris Pismenny <borisp@mellanox.com>
Signed-off-by: Tariq Toukan <tariqt@mellanox.com>
Signed-off-by: Raed Salem <raeds@mellanox.com>
Signed-off-by: Huy Nguyen <huyn@mellanox.com>