optee: model OP-TEE as a platform device/driver
To simplify adding ACPI support to the OP-TEE driver, model it as
a platform driver. This will permit us to use the generic device
property layer for parsing additional properties, regardless of
whether DT or ACPI is being used.

Note that this change will result in the OP-TEE driver to be loaded
automatically on systems that advertise the presence of OP-TEE via
the device tree.

Change-Id: If1291b627c44b8a41b7f7a22b9d0a20b6d0e8da6
Signed-off-by: Ard Biesheuvel <ardb@kernel.org>