Every IP layer receive handler has a check that if a receive skb matches an existing IPsec policy, the skb must be stamped decrypted. The check includes checking if the sec_path is set and the related SA matches the skb's src and dst IP address.
For "full offload" packet, the network stack will not see ESP packet.
Therefore, there is no sec_path set for the skb and no evidence to show
that the packet that matches the IPsec policy is actually decrypted. In this case,
instead of drop the packet, search for the matching "full offload" SAs in the
database, if found, allow the packet.

This change is mainly for the "first new packet" of the OVS before OVS offload
the rule into Hardware. So performance concern because of querying SADB should be
negligible.

source VM send normal packet --> Hypervisor --> OVS --> VXLAN encapsulation
--> xfrm_out, do nothing here (TX full offload datapath patch)
--> IPsec full offload (encrypt, encap) --> out of the wire
--> reach other side --> IPsec full offload (decap, decrypt) --> OVS
--> xfrm_check protocol's receive (scan SADB for full offload to accept the packet)
--> VXLAN decapsulation --> OVS --> destination VM

Signed-off-by: Raed Salem <raeds@mellanox.com>