Viper: Improved workaround for lost MSI-X bug
In BlueField-1, all transactions coming out of the GIC were
forced to be coherent by the skymesh interface.  That behavior
was changed in BlueField-2 so that the skymesh honors the
cacheability attributes of the requests coming in from the
GIC AXI4 bus.  For reasons we do not yet understand,
non-cacheable accesses do not work reliably.

In addition to that, the GIC-500 does not advertise its
support for cacheable access in the usual way as specified
by the GICv3 architecture, which is through the Shareability
fields of the various BASE registers.  However, it is possible
to force it to use cacheable accesses by ignoring the value
of Shareabilty field and setting the InnerShareable field
directly.

Change-Id: Ia22e6259772c491b879685b7d719fd7e7eef6192
Depends-On: ezchip~bluefield-rel/3.0~Ia783371584d3bd96b0fa8b8d44c1e0ecbb5b127d
Depends-On: linux~bfdev-5.4.31~Ia22e6259772c491b879685b7d719fd7e7eef6192