mlxbf-gige: switch OOB PHY logic to use polling mode
This delivery changes the OOB driver logic to specify
that the OOB PHY device should be polled by the Linux
PHY framework instead of being interrupt-driven.  The
interrupt-driven design can be re-visited in a future
patch. Also, as a simplification, the call to PHY API
"phy_start_aneg()" was observed to be unnecessary and
thus was removed.  This API call is instead handled
by the Linux PHY state machine when needed.

The driver version is bumped to 1.1 to signify this
change in OOB PHY behavior.

RM #2196423

Change-Id: I8537b5f10d666992ffd76b7ae9ed24b5a34586d9