mlxbf-gige: call napi_schedule after TX ring full detected
This delivery fixes an issue with the driver transmit logic
that can cause a transmit timeout.  During periods of heavy
traffic, the driver's TX ring can fill up.  When the TX ring
fills up the TX queue of the network interface is stopped.
If 5 seconds elapse while in the "TX queue stopped" state,
then the kernel throws the following error:
   NETDEV WATCHDOG: oob_net0 (mlxbf_gige): transmit queue 0 timed out
The fix for this is to add an explicit call to napi_schedule()
when the driver TX ring is detected to be full. There is no
separate "TX complete" interrupt, that is why napi_schedule()
needs to be manually called.  This driver's NAPI poll routine
will process TX completions, and will hopefully drain the TX
ring allowing the netif TX queue to be awakened.

This transmit timeout issue was most likely reproduced if
the driver rings were set to "rx 64 tx 64" by "ethtool -G".
The "iperf3" test tool used to transmit packets to the OOB
interface.

RM #2192019

Change-Id: I219c13af714098d148c7aa4b498d77bdd8c60451