tmfifo: Fix a memory barrier issue
The virtio framework uses wmb() when updating avail->idx. It
gurantees the write order, but not necessarily loading order
for the code accessing the memory. This commit adds a load barrier
after reading the avail->idx to make sure all the data in the
descriptor is visible. It also adds a barrier when returning the
packet to virtio framework to make sure read/writes are visible to
the virtio code.

RM #2269638

Change-Id: I0759e75a8863b3cee9eb0b3794cd87c113019f72