Asgard bringup: Fix OOB PHY interrupt
During Asgard bringup (FHHL controller), we noticed
that the oob interface does not come up.
At boot time, there are 2 PHY interrupt triggered:
the first one signals link down, and the second one
signals a link up.
What happens on the Asgard: the link down interrupt
is triggered but not the link up.
The PHY device behaves slightly different on this
specific board but it is the same issue as we have
seen before.
The linux opensource code triggers a Software PHY
reset in several instances, which wipes out the
interrupt controller/status register (0x1B).
When it does that, no more interrupts are received
on the PHY.
So we need to always make sure to reenable the PHY
interrupt enable bits for link up/down.

RM #2356230

Change-Id: I0dfd9918c410d5889b09805f1c4e533bc2176ef0