net/mlx5: Eswitch/IPsec: Create Tx IPsec full offload tables esw
When enable SRIOV switchdev mode, if IPsec mode is full, create
Tx IPsec FDB table. This table acts as a termination table for
all the packets whose destination is uplink vport. In another word,
all the uplink vport transmit steering rules in tc FDB and
slow patch FDB will be forwarded to the Tx IPsec FDB.
If the match criteria is met, packet will be encap and encrypted
and forwarded to uplink vport. On any error, the packet is dropped.
Flow counters are supported for both good and error flows.

Issue: 2194228
Change-Id: I4bdcccc4b6827b435d32f0115f5da93b8d81c1f6
Signed-off-by: Raed Salem <raeds@nvidia.com>
Signed-off-by: Huy Nguyen <huyn@nvidia.com>