Race condition in the GPIO and OOB interrupts
The OOB and GPIO drivers enable the YU GPIO interrupts
before registering their respective interrupt handlers.

Because of the above, the interrupt could be triggered
before the interrupt handler is registered. And so it
would just be missed, and the interrupt registers are
never cleared and prevent any further interrupts.

Even with the fix above, we still see occasionally
(during reboot or powercycle) a race condition in
the OOB driver between the time the YU GPIO PHY
interrupt is enabled and the time the PHY interrupt
is enabled. To make we dont fall in this case,
several, we can just retrigger autonegotiation
once both mlxbf_gige_mdio_enable_phy_int and
mlxbf_gige_phy_enable_interrupt have been
called.

Several other open source drivers do that as well.
For instance: lan743x_main.c

Another fix that was made in this patch is restricting
the support of the AUX power mode interrupt only
to OCP3.0 board.

RM #2356230

Change-Id: I3b82526815a655cd59046a759eee54ca5d056d2e