Fix PHY KSZ9031 interrupt issue
In high and low temperature conditions (+60 and -20 degrees Celsius),
the OOB interface doesn't always come up after powercycling or after
a software reset.
This is due to the PHY interrupt control/status register 0x1B being
cleared.

That register is cleared because the micrel.c code triggers a soft
reset of the PHY.

The PHY device has a HW bug where the KSZ9031 will appear to complete
autonegotiation, but will drop all traffic afterwards. When this
happens, the idle error count will read 0xFF after autonegotiation
completes. So the micrel.c driver resets the PHY when in that state.

When that reset happens, the interrupt register is cleared so it
needs to be set again to its previous state.

The issue encountered here is a race condition between setting
the interrupt register again and the completion of the reset.
To avoid this race condition, a delay was added to give ample
time for the reset to complete.

The idle error count reaches its max value due to the very high
(or low) temperatures. Although the room temperature of the
BlueField-2 system is 60C, the RJ45 junction of the OOB is 70C.
According to the manufacturer, the temperature limit for the
RJ45 connector is 70C. The HW team will fix this but in the mean
time this SW workaround is needed.

Change-Id: I45780b31afc38864c4da89a99dca84755ba17a18