OOB PHY workaround and remove previous commit for micrel.c
The micrel.c change is no longer needed.

The current implementation of the PHY interrupt handler
is internal to the MAC driver. So the phydev->irq variable
is set to PHY_IGNORE_INTERRUPT. This conflicts with the way
the micrel.c driver was designed. If irq=PHY_IGNORE_INTERRUPT,
none of the code enabling/disabling the PHY interrupt is
executed. To workaround that, the irq is set only before the
phy_start routine is called. At the moment, it shouldn't be
called before phy_connect_direct because we are not using the
phy.c interrupt handler.

Change-Id: I49f229cd4aed11e09488c96414113a86e248581d