BlueSphere crash due to ipmb-host
On the Bluesphere (only after powercycling not SW reset), the kernel
crashes and the UART is inaccessible and the user is unable to ssh
through TMFIFO or OOB interface.

At boot time, the ipmb-host.c driver tries to execute a handshake with
the BMC. Since the openbmc dgx image takes longer to boot than the
BlueField, that handshake fails because the BMC IPMI service are not up
yet.

There are 2 bugs responsible for this:
1) use spin_lock_irqsave/spin_unlock_irqrestore instead of
spin_lock_irq/spin_unlock_irq.

2) the msg_to_send variable shouldn't be set to NULL so early because
the ipmb_send function keeps sending IPMI request (passed by
ipmi_msghandler.c) and this causes issues with the "wait_event" function
called from __get_device_id.

RM #2404917

Change-Id: I9e4bcb2688b690da2a3aa93cae15e9f705f09677