Revert "netfilter: nf_flow_table_offload: Set NF_FLOW_HW if offload work queued"
This reverts commit de81e850f06e16fc53c374b7651c17933941c71c.

Before commit e7f6c451b7cd ("netfilter: flowtable: Do flow refresh when
requested") we only did refresh if add work queued and the nf flow bit
was set. After it, we could do refresh if if the hw offload bit was not
set. The reverted commit was some kind of optimization that is not actually
needed. Revert it so gc_step won't think flow is not in hw and free the
flow while refresh could happen.

Issue: 2454882
Fixes: e7f6c451b7cd ("netfilter: flowtable: Do flow refresh when requested")
Change-Id: Ia9cb5abe3f331c2d783dc447242f763e7eeec12e
Signed-off-by: Roi Dayan <roid@nvidia.com>