mlxbf-gige: fix start_xmit return status on errors
This driver delivery fixes the status that is returned
in the error paths of the "start_xmit" method.  The
previous value was NET_XMIT_DROP, but this value is
not part of the "netdev_tx_t" enum, which is the return
type of the "start_xmit" method.  Instead the value
returned in these error paths should be NETDEV_TX_OK,
since the driver "took care of the packet", as described
in include/linux/netdevice.h

This issue was flagged by Coverity.

RM #2352465

Change-Id: Ib8bbc81ba989b7528f393893123688bcb75a1ba4