mlxbf_gige: cleanup support for ARM32 build
As part of upstreaming this GIGE driver, a request was made to
allow it to be built for the ARM32 architecture.  When the module
main.c was broken into multiple modules, the support for the ARM32
build was lost.  This delivery introduces a set of changes that
re-enables the ARM32 build and cleans up some typecast issues.
The header file "io-64-nonatomic-lo-hi.h" provides APIs for supporting
64-bit readq/writeq operations in a non-atomic fashion on the ARM32
architecture.

Change-Id: I67c8a40c01b5bb748ba2cb1cce3e89ef4099beed